requirements.txt
Flask==2.2.5
Markdown==3.6
openai==1.37.1
pdfminer.six==20240706
python-dotenv==1.0.1
Requests==2.32.3
gunicorn

